refactor(utils): Reorganise to simplify imports
