--- lib/utils/isLight.ts
@@ -1,3 +1,0 @@
-import { getLuminance } from 'polished';
-
-export default (color: string) => getLuminance(color) > 0.6;
