+++ lib/utils/isLight/index.ts
@@ -1,0 +1,3 @@
+import { getLuminance } from 'polished';
+
+export const isLight = (color: string) => getLuminance(color) > 0.6;
