+++ lib/utils/mapToStyleProperty/index.ts
@@ -2,7 +2,10 @@ import mapValues from 'lodash/mapValues';
 import { Properties } from 'csstype';
 import { Styles } from 'sku/treat';
 
-export default <Map extends string, Value extends string | number>(
+export const mapToStyleProperty = <
+  Map extends string,
+  Value extends string | number
+>(
   map: Record<Map, Value>,
   propertyName: keyof Properties,
   mapper?: (value: Value, propertyName: keyof Properties) => Styles,
