--- lib/utils/toUnit.ts
@@ -1,1 +1,0 @@
-export const px = (num: number) => `${num}${num !== 0 ? 'px' : ''}`;
